Delete Fx stack
