--- module/ldbc-net/js/src/main/scala/ldbc/net/effect/PlatformTls.scala
@@ -1,107 +1,0 @@
-/**
- * Copyright (c) 2023-2026 by Takahiko Tominaga
- * This software is licensed under the MIT License (MIT).
- * For more information see LICENSE or https://opensource.org/licenses/MIT
- */
-
-package ldbc.net.effect
-
-import java.util.concurrent.atomic.AtomicBoolean
-
-import scala.scalajs.js
-import scala.scalajs.js.typedarray.Uint8Array
-import scala.util.control.NonFatal
-
-import ldbc.effect.Async
-import ldbc.net.{ HostnameMatcher, NodeRawSocket, SSL, TrustSource }
-
-/**
- * Scala.js generic TLS engine over node's `tls.connect` (design Phase 2), generic over the effect `F`. The
- * trust / identity / protocol-version option mapping is identical to `ldbc.net.PlatformTls`: the plaintext
- * [[NodeRawSocket]] is detached (listeners removed, pre-read bytes `unshift`ed back onto the stream) and the
- * underlying node socket is handed to `tls.connect`, which performs the handshake, chain validation
- * (`rejectUnauthorized`) and — via `servername`/`host` — node's built-in hostname verification. The resulting
- * `TLSSocket` is wrapped in a fresh [[NodeRawSocket]], so reads and writes carry plaintext.
- */
-private[net] object PlatformTls:
-
-  private lazy val tlsModule = js.Dynamic.global.require("tls")
-
-  /** Wraps `socket` in a TLS client session. */
-  def client[F[_]](socket: Socket[F], host: String, port: Int, ssl: SSL)(using F: Async[F]): F[Socket[F]] =
-    val _ = port
-    socket match
-      case backed: RawBackedSocket =>
-        backed.underlying match
-          case node: NodeRawSocket => upgrade(node, host, ssl)
-          case _                   =>
-            F.raiseError(new IllegalArgumentException("JS TLS requires a socket produced by the node IoEngine"))
-      case _ =>
-        F.raiseError(new IllegalArgumentException("JS TLS requires a socket produced by the node IoEngine"))
-
-  /** Detaches the plaintext wrapper, replays pre-read bytes, and starts the TLS handshake. */
-  private def upgrade[F[_]](node: NodeRawSocket, host: String, ssl: SSL)(using F: Async[F]): F[Socket[F]] =
-    F.async { cb =>
-      F.delay {
-        val done = new AtomicBoolean(false)
-        try
-          val (raw, pending) = node.detachForUpgrade()
-          if pending.nonEmpty then raw.unshift(toUint8Array(pending))
-          val options = buildOptions(host, ssl)
-          options.socket = raw
-          val tlsSock = tlsModule.connect(options)
-          tlsSock.on(
-            "secureConnect",
-            ((
-              () => if done.compareAndSet(false, true) then cb(Right(Socket.fromRaw[F](new NodeRawSocket(tlsSock))))
-            ): js.Function0[Unit])
-          )
-          tlsSock.on(
-            "error",
-            (
-              (err: js.Dynamic) =>
-                if done.compareAndSet(false, true) then cb(Left(new RuntimeException(s"TLS handshake failed: $err")))
-            ): js.Function1[js.Dynamic, Unit]
-          )
-          Some(F.delay { tlsSock.destroy(); () }): Option[F[Unit]]
-        catch
-          case NonFatal(error) =>
-            if done.compareAndSet(false, true) then cb(Left(error))
-            Option.empty[F[Unit]]
-      }
-    }
-
-  /** Maps a [[SSL]] to node `tls.connect` options (trust, identity, protocol versions). */
-  private def buildOptions(host: String, ssl: SSL): js.Dynamic =
-    val options = js.Dynamic.literal()
-    ssl match
-      case SSL.System                                                 => ()
-      case SSL.Trusted                                                => options.rejectUnauthorized = false
-      case SSL.Custom(trust, clientAuth, verifyHostname, tlsVersions) =>
-        if clientAuth.isDefined then
-          throw new UnsupportedOperationException("client certificates (mTLS) are not yet supported")
-        trust match
-          case TrustSource.System            => ()
-          case TrustSource.InsecureTrustAll  => options.rejectUnauthorized = false
-          case TrustSource.FromPemCerts(pem) => options.ca                 = pem
-        if !verifyHostname then
-          options.checkServerIdentity =
-            ((_: js.Dynamic, _: js.Dynamic) => js.undefined): js.Function2[js.Dynamic, js.Dynamic, js.Any]
-        if tlsVersions.nonEmpty then
-          options.minVersion = tlsVersions.min
-          options.maxVersion = tlsVersions.max
-      case SSL.None           => ()
-      case SSL.Platform(_, _) =>
-        throw new UnsupportedOperationException("SSL.Platform (a JVM SSLContext) is not supported on Scala.js")
-    options.host                                                  = host
-    if !HostnameMatcher.isIpLiteral(host) then options.servername = host
-    options
-
-  /** Converts bytes to a `Uint8Array` for `stream.unshift`. */
-  private def toUint8Array(bytes: Array[Byte]): Uint8Array =
-    val out = new Uint8Array(bytes.length)
-    var i   = 0
-    while i < bytes.length do
-      out(i) = (bytes(i) & 0xff).toShort
-      i += 1
-    out
