+++ module/ldbc-net/jvm/src/test/scala/ldbc/net/GenericSocketFxTest.scala
@@ -6,8 +6,8 @@
 
 package ldbc.net
 
-import java.util.concurrent.atomic.AtomicReference
 import java.util.concurrent.CountDownLatch
+import java.util.concurrent.atomic.AtomicReference
 
 import scala.concurrent.duration.*
 
@@ -15,10 +15,11 @@ import munit.FunSuite
 
 import ldbc.fx.{ Fx, FxRuntime }
 import ldbc.fx.concurrentFx // given Concurrent[Fx]
-import ldbc.net.effect.IoEngine
+
+import ldbc.net.IoEngine
 
 /**
- * Step 3 check: the generic `ldbc.net.effect.Socket[F]` over the effect-free [[NioRawEngine]] runs at
+ * Step 3 check: the generic `ldbc.net.Socket[F]` over the effect-free [[NioRawEngine]] runs at
  * `F = Fx` (via `concurrentFx`) against a real MySQL server — no `Fx`-native `Socket` involved.
  */
 class GenericSocketFxTest extends FunSuite:
@@ -31,8 +32,8 @@ class GenericSocketFxTest extends FunSuite:
     ref.get().fold(throw _, identity)
 
   test("generic Socket[Fx] over NioRawEngine reads the MySQL handshake (proto v10)") {
-    val engine:  IoEngine[Fx] = IoEngine.fromRaw[Fx](NioRawEngine.global)
-    val program: Fx[Byte]     =
+    val engine: IoEngine[Fx] = IoEngine.fromRaw[Fx](NioRawEngine.global)
+    val program: Fx[Byte] =
       engine.connect("127.0.0.1", 13306, 5.seconds).flatMap { sock =>
         sock.read(4).flatMap {
           case Some(header) =>
